requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pillow>=10.4.0
numpy>=2.1.0
